requirements.txt
email-validator>=2.2.0
flask>=3.0.3
flask-sqlalchemy>=3.1.1
stripe>=11.1.0
pillow>=10.4.0
sqlalchemy>=2.0.35
werkzeug>=3.0.4
requests>=2.32.3
openai>=1.51.2
flask-socketio>=5.3.6
gradio>=5.1.0
python-dotenv>=1.0.0